fix: unify anglemod to C-parity 16-bit quantization
Consolidate all anglemod implementations to use the canonical
pkg/types.AngleMod with C Ironwail's 16-bit quantization (65536 steps):

- pkg/types: fix AngleMod from 8-bit (256 steps) to 16-bit (65536 steps)
- internal/client/relink.go: replace continuous math.Mod with types.AngleMod
- internal/server/movement.go: remove duplicate anglemod, use types.AngleMod
- pkg/types: add test verifying quantization matches C behavior

Fixes: ironwail-go-1b4o

+++ pkg/types/types.go
@@ -215,7 +215,7 @@ func ClampInt(value, min, max int) int {
 // This uses 16-bit precision (~0.0055° per step), matching the exact
 // quantization behavior of the C Quake engine.
 func AngleMod(angle float32) float32 {
-	return (360.0 / 65536) * float32(int(angle*(65536.0/360.0))&65535)
+	return float32((360.0 / 65536.0) * float64(int(float64(angle)*(65536.0/360.0))&65535))
 }
 
 // AngleByte converts a float angle (in degrees) to the 8-bit byte
